Replace zip-local package with adm-zip
Background
==========
Viperlight scan in the feature pipeline found a package using a version with a security vulnerability (jzip).
The zip-local package used by this solution has a dependency on jzip, but it has not been updated to the latest version of jzip.

Changes
=======
Replaced zip-local package and used adm-zip instead.
- updated package.json and package-lock.json
- updated lambda code that was using zip-local
- updated lambda layer directory to replace zip-local with adm-zip
- updated code referencing the lambda layer directory
- updated install_dependencies scripts