Remove function to verify ipv6
Background
==========
SonarQube scan created a code smell issue for the validateIPv4 function for its overly complex regex string. To address this code smell, the function was removed and its usage was replaced with ```net.isIPv4()``` built in function from node.
For consistency and best practices, the validateIPv6 function will also be removed and replaced with ```net.isIPv6()```

Changes
=======
- removed validateIPv6 function and replaced its usage with net.isIPv6
  - also removed its unit tests
- simplified calls to ```.some()```