Updated versioning method for the solution
Solution versioning info
========================
Originally, the solution's version was hard-coded in the main cdk app file: source/bin/secure_media_stream.ts
Later, this hard-coded value was updated so that the version will dynamically update by replacing the %%VERSION%% string when executing the deployment/build-s3-dist.sh script. This change, however, introduced an issue when deploying the solution using cdk commands (https://t.corp.amazon.com/D95944302/overview).
So instead of replacing the %%VERSION%% cdk.context will be used to handle the solution's versioning. Additionally, the deployment/build-s3-dist.sh logic will change so that it passes the appropriate version when executing cdk synth via the --context variable
Finally, the script also replaces the %%BUCKET_NAME%% and %%SOLUTION_NAME%% string, but it has discovered that these string do not exist anywhere in the template so there was no need to replace them.

CodeBuild npm install info
==========================

In the buildspec.yml pre_build commands, it instructed CodeBuild to update npm to the latest version
before running the Build. The issue with this is that there are cases when the latest npm version
is not supported with the node runtime which causes the build to fail.

Changes
=======

- deployment/build-s3-dist.sh updated to use context parameter to version the solution instead of %%VERSION%% string replacement. Also, removed bucket_name and solution_name string replacement commands as they are not necessary
- source/bin/secure_media_stream.ts updated to read context from cdk.context.json file
- deleted npm install npm@latest command from builspec.yml
- Added unit tests to ensure cdk.json and package.json versions match